+++ web/src/components/Navigation/Navigation.structure.tsx
@@ -3,7 +3,7 @@ import { Outlet, useNavigate } from 'react-router-dom'
 import { useUserContext } from '../../utils/hooks/useUserContext'
 import { useUserLogInStatus } from '../../utils/hooks/useUserLogInStatus'
 import { IUser } from '../../utils/interfaces/interfaces'
-import Login from '../User/LogIn/LogIn'
+import Login from '../../pages/LogIn/LogIn'
 import Background from './Background/Background.component'
 import Banner from './Banner.component'
 import NavBar from './NavBar/NavBar.component'
