--- web/src/components/User/UserPortal/UserPortal.tsx
@@ -1,9 +1,0 @@
-import "./UserPortal.css";
-const UserPortal = () => {
-  return (
-    <div>
-      <h1>FIXME</h1>
-    </div>
-  );
-};
-export default UserPortal;
